added border color and move button files
