+++ src/scripts/components/buttons/loginButton.js
@@ -1,4 +1,4 @@
-import signIn from '../helpers/signIn';
+import signIn from '../../helpers/signIn';
 
 // GOOGLE LOGIN BUTTON
 const loginButton = () => {
