+++ src/scripts/components/domBuilder.js
@@ -2,6 +2,7 @@ const domBuilder = () => {
   document.querySelector('#app').innerHTML = `
     <div id="navigation"></div>
     <div id="main-container">
+        <div id="lt"></div>
         <div id="add-button"></div>
         <div id="form-container"></div>
         <div id="store"></div>
